fix: use useNavigate to avoid style issue
